added fixed point vorbis library tremor

+++ mp3decoders/src/main/java/com/axelby/mp3decoders/MainActivity.java
@@ -5,7 +5,6 @@ import android.content.Context;
 import android.media.AudioFormat;
 import android.media.AudioManager;
 import android.media.AudioTrack;
-import android.media.MediaPlayer;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.View;
@@ -13,6 +12,7 @@ import android.widget.CheckBox;
 import android.widget.RadioGroup;
 import android.widget.TextView;
 
+import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
@@ -28,11 +28,71 @@ import javazoom.jl.decoder.SampleBuffer;
 public class MainActivity extends Activity {
 	private TextView _stateText;
 	private AudioTrack _track = null;
+	IMediaDecoder _decoder;
+
+	private Runnable vorbisRunnable = new Runnable() {
+		@Override
+		public void run() {
+			Vorbis vorbis = new Vorbis(getFilesDir() + "/loop1_ogg.ogg");
+			_decoder = vorbis;
+
+			_track = new AudioTrack(AudioManager.STREAM_MUSIC,
+					vorbis.getRate(),
+					vorbis.getNumChannels() == 2 ? AudioFormat.CHANNEL_OUT_STEREO : AudioFormat.CHANNEL_OUT_MONO,
+					AudioFormat.ENCODING_PCM_16BIT,
+					vorbis.getRate() * 2,
+					AudioTrack.MODE_STREAM);
+			_track.setPositionNotificationPeriod(vorbis.getRate());
+			RadioGroup rg = (RadioGroup) findViewById(R.id.playbackRate);
+			switch (rg.getCheckedRadioButtonId()) {
+				case R.id.rate15:
+					_track.setPlaybackRate((int) (vorbis.getRate() * 1.5));
+					break;
+				case R.id.rate20:
+					_track.setPlaybackRate((int) (vorbis.getRate() * 2.0));
+					break;
+			}
+			_track.setPlaybackPositionUpdateListener(playbackPositionListener);
+
+
+			_track.play();
+			changeState("playing");
+
+			try {
+				int total = 0;
+				long start = System.currentTimeMillis();
+				short[] pcm = new short[1024 * 5];
+				while (vorbis.readSamples(pcm, 0, pcm.length) > 0) {
+					total += pcm.length;
+					if (_track.getPlayState() != AudioTrack.PLAYSTATE_PLAYING) {
+						Thread.sleep(50);
+						continue;
+					}
+					if (!isTimingOnly()) {
+						_track.write(pcm, 0, pcm.length);
+						//Log.i("mp3decoders", "wrote " + (pcm.length / 2) + " frames");
+					}
+				}
+				long end = System.currentTimeMillis();
+				Log.i("mp3decoders", "vorbis decoded " + total + " frames in " + (end - start) + " milliseconds");
+
+			} catch (InterruptedException e) {
+				Log.e("mp3decoders", "InterruptedException", e);
+			} finally {
+				vorbis.close();
+				waitAndCloseTrack();
+			}
+
+			Log.i("mp3decoders", "done loading audiotrack");
+			changeState("finished playing");
+		}
+	};
 
 	private Runnable mpg123Runnable = new Runnable() {
 		@Override
 		public void run() {
 			MPG123 mpg123 = new MPG123(getFilesDir() + "/loop1.mp3");
+			_decoder = mpg123;
 
 			_track = new AudioTrack(AudioManager.STREAM_MUSIC,
 					mpg123.getRate(),
@@ -40,16 +100,7 @@ public class MainActivity extends Activity {
 					AudioFormat.ENCODING_PCM_16BIT,
 					mpg123.getRate() * 2,
 					AudioTrack.MODE_STREAM);
-			double secs = 1.0 * mpg123.getLength() / mpg123.getRate();
-			Log.i("jlayerplayer", "secs: " + secs);
-			Log.i("jlayerplayer", "secs per frame: " + mpg123.getSecondsPerFrame());
-			Log.i("jlayerplayer", "length: " + mpg123.getLength());
-			Log.i("jlayerplayer", "secs per frame * length: " + mpg123.getSecondsPerFrame() * mpg123.getLength());
-
-			MediaPlayer mp = MediaPlayer.create(MainActivity.this, R.raw.loop1);
-			Log.i("jlayerplayer", "mediaplayer duration: " + mp.getDuration());
-
-			_track.setPositionNotificationPeriod((int) (1000.0f * 1000.0f * mpg123.getSecondsPerFrame()));
+			_track.setPositionNotificationPeriod(mpg123.getRate());
 			RadioGroup rg = (RadioGroup) findViewById(R.id.playbackRate);
 			switch (rg.getCheckedRadioButtonId()) {
 				case R.id.rate15:
@@ -66,29 +117,31 @@ public class MainActivity extends Activity {
 			changeState("playing");
 
 			try {
+				int total = 0;
 				long start = System.currentTimeMillis();
 				short[] pcm = new short[1024 * 5];
 				while (mpg123.readSamples(pcm, 0, pcm.length) > 0) {
+					total += pcm.length;
 					if (_track.getPlayState() != AudioTrack.PLAYSTATE_PLAYING) {
 						Thread.sleep(50);
 						continue;
 					}
 					if (!isTimingOnly()) {
 						_track.write(pcm, 0, pcm.length);
-						Log.i("jlayerplayer", "wrote " + (pcm.length / 2) + " frames");
+						//Log.i("mp3decoders", "wrote " + (pcm.length / 2) + " frames");
 					}
 				}
 				long end = System.currentTimeMillis();
-				Log.i("jlayerplayer", "jlayer decoded in " + (end - start) + " milliseconds");
+				Log.i("mp3decoders", "mpg123 decoded " + total + " frames in " + (end - start) + " milliseconds");
 
 			} catch (InterruptedException e) {
-				Log.e("jlayerplayer", "InterruptedException", e);
+				Log.e("mp3decoders", "InterruptedException", e);
 			} finally {
 				mpg123.close();
 				waitAndCloseTrack();
 			}
 
-			Log.i("jlayerplayer", "done loading audiotrack");
+			Log.i("mp3decoders", "done loading audiotrack");
 			changeState("finished playing");
 		}
 	};
@@ -98,7 +151,7 @@ public class MainActivity extends Activity {
 		public void run() {
 			InputStream loop1 = null;
 			try {
-				Log.i("jlayerplayer", "in loader thread");
+				Log.i("mp3decoders", "in loader thread");
 
 				loop1 = getResources().openRawResource(R.raw.loop1);
 				Decoder decoder = new Decoder();
@@ -123,7 +176,7 @@ public class MainActivity extends Activity {
 
 					if (!isTimingOnly()) {
 						_track.write(pcm, 0, pcm.length);
-						Log.i("jlayerplayer", "wrote " + (pcm.length / 2) + " frames");
+						//Log.i("mp3decoders", "wrote " + (pcm.length / 2) + " frames");
 					}
 
 					header = bitstream.readFrame();
@@ -131,22 +184,22 @@ public class MainActivity extends Activity {
 						done = true;
 				}
 				long end = System.currentTimeMillis();
-				Log.i("jlayerplayer", "jlayer decoded in " + (end - start) + " milliseconds");
+				Log.i("mp3decoders", "jlayer decoded in " + (end - start) + " milliseconds");
 
-				Log.i("jlayerplayer", "done loading audiotrack");
+				Log.i("mp3decoders", "done loading audiotrack");
 				changeState("finished playing");
 			} catch (BitstreamException e) {
-				Log.e("jlayerplayer", "bitstreamexception", e);
+				Log.e("mp3decoders", "bitstreamexception", e);
 			} catch (DecoderException e) {
-				Log.e("jlayerplayer", "decoderexception", e);
+				Log.e("mp3decoders", "decoderexception", e);
 			} catch (InterruptedException e) {
-				Log.e("jlayerplayer", "InterruptedException", e);
+				Log.e("mp3decoders", "InterruptedException", e);
 			} finally {
 				try {
 					if (loop1 != null)
 						loop1.close();
 				} catch (IOException e) {
-					Log.e("jlayerplayer", "ioexception", e);
+					Log.e("mp3decoders", "ioexception", e);
 				}
 
 				waitAndCloseTrack();
@@ -167,11 +220,12 @@ public class MainActivity extends Activity {
 				while (_track.getPlaybackHeadPosition() != 0)
 					Thread.sleep(10);
 			} catch (InterruptedException e) {
-				Log.e("jlayerplayer", "InterruptedException", e);
+				Log.e("mp3decoders", "InterruptedException", e);
 			}
 
 			_track.release();
 			_track = null;
+			_decoder = null;
 		}
 	}
 
@@ -182,7 +236,7 @@ public class MainActivity extends Activity {
 				AudioFormat.ENCODING_PCM_16BIT,
 				header.bitrate() * 2,
 				AudioTrack.MODE_STREAM);
-		Log.i("jlayerplayer", "ms per frame: " + header.ms_per_frame());
+		Log.i("mp3decoders", "ms per frame: " + header.ms_per_frame());
 		track.setPositionNotificationPeriod((int) (1000.0f * header.ms_per_frame()));
 		RadioGroup rg = (RadioGroup) findViewById(R.id.playbackRate);
 		switch (rg.getCheckedRadioButtonId()) {
@@ -205,14 +259,18 @@ public class MainActivity extends Activity {
 
 		@Override
 		public void onPeriodicNotification(AudioTrack audioTrack) {
-			changeState("periodic notification at " + audioTrack.getPlaybackHeadPosition());
+			if (_decoder != null)
+				changeState(String.format("periodic notification at %.2f, head position %d",
+						_decoder.getPosition(), audioTrack.getPlaybackHeadPosition()));
+			else
+				changeState("periodic notification at " + audioTrack.getPlaybackHeadPosition());
 		}
 	};
 
 	private AudioManager.OnAudioFocusChangeListener audioFocusListener = new AudioManager.OnAudioFocusChangeListener() {
 		@Override
 		public void onAudioFocusChange(int state) {
-			Log.i("jlayerplayer", "audiofocus state " + state);
+			Log.i("mp3decoders", "audiofocus state " + state);
 			if (_track == null)
 				return;
 
@@ -242,7 +300,7 @@ public class MainActivity extends Activity {
 		public void onClick(View view) {
 			if (_track != null) {
 				_track.pause();
-				Log.i("jlayerplayer", "paused");
+				Log.i("mp3decoders", "paused");
 			}
 		}
 	};
@@ -252,14 +310,14 @@ public class MainActivity extends Activity {
 		public void onClick(View view) {
 			if (_track != null) {
 				_track.play();
-				Log.i("jlayerplayer", "resumed");
+				Log.i("mp3decoders", "resumed");
 				return;
 			}
 
 			if (!requestAudioFocus())
 				return;
 			new Thread(jLayerRunnable).start();
-			Log.i("jlayerplayer", "started JLayer thread");
+			Log.i("mp3decoders", "started JLayer thread");
 		}
 	};
 
@@ -268,27 +326,42 @@ public class MainActivity extends Activity {
 		public void onClick(View view) {
 			if (_track != null) {
 				_track.play();
-				Log.i("jlayerplayer", "resumed");
+				Log.i("mp3decoders", "resumed");
 				return;
 			}
 
 			if (!requestAudioFocus())
 				return;
 			new Thread(mpg123Runnable).start();
-			Log.i("jlayerplayer", "started MPG123 thread");
+			Log.i("mp3decoders", "started MPG123 thread");
 		}
 	};
 
+	private View.OnClickListener playVorbisHandler = new View.OnClickListener() {
+		@Override
+		public void onClick(View view) {
+			if (_track != null) {
+				_track.play();
+				Log.i("mp3decoders", "resumed");
+				return;
+			}
+
+			if (!requestAudioFocus())
+				return;
+			new Thread(vorbisRunnable).start();
+			Log.i("mp3decoders", "started Vorbis thread");
+		}
+	};
 	private boolean requestAudioFocus() {
 		AudioManager audioManager = (AudioManager) getSystemService(Context.AUDIO_SERVICE);
 		int result = audioManager.requestAudioFocus(audioFocusListener,
 				AudioManager.STREAM_MUSIC,
 				AudioManager.AUDIOFOCUS_GAIN);
 		if(result != AudioManager.AUDIOFOCUS_REQUEST_GRANTED) {
-			Log.d("jlayerplayer", "Can't get audio focus");
+			Log.d("mp3decoders", "Can't get audio focus");
 			return false;
 		}
-		Log.i("jlayerplayer", "audiofocus request granted");
+		Log.i("mp3decoders", "audiofocus request granted");
 		return true;
 	}
 
@@ -299,22 +372,41 @@ public class MainActivity extends Activity {
 
 		findViewById(R.id.playJLayer).setOnClickListener(playJLayerHandler);
 		findViewById(R.id.playMPG123).setOnClickListener(playMPG123Handler);
+		findViewById(R.id.playVorbis).setOnClickListener(playVorbisHandler);
 		findViewById(R.id.pause).setOnClickListener(pauseHandler);
 		_stateText = (TextView) findViewById(R.id.state);
 		changeState("init");
 
 		try {
-			InputStream loop1 = getResources().openRawResource(R.raw.loop1);
-			FileOutputStream out = new FileOutputStream(getFilesDir() + "/loop1.mp3");
-			byte[] buffer = new byte[1024];
-			int len;
-			while ((len = loop1.read(buffer)) != -1)
-				out.write(buffer, 0, len);
-			loop1.close();
+			if (!new File(getFilesDir() + "/loop1.mp3").exists()) {
+				InputStream loop1 = getResources().openRawResource(R.raw.loop1);
+				FileOutputStream out = new FileOutputStream(getFilesDir() + "/loop1.mp3");
+				byte[] buffer = new byte[1024];
+				int len;
+				while ((len = loop1.read(buffer)) != -1)
+					out.write(buffer, 0, len);
+				loop1.close();
+			}
 		} catch (FileNotFoundException e) {
-			Log.e("jlayerplayer", "FileNotFoundException", e);
+			Log.e("mp3decoders", "FileNotFoundException", e);
 		} catch (IOException e) {
-			Log.e("jlayerplayer", "IOException", e);
+			Log.e("mp3decoders", "IOException", e);
+		}
+
+		try {
+			if (!new File(getFilesDir() + "/loop1_ogg.ogg").exists()) {
+				InputStream loop1 = getResources().openRawResource(R.raw.loop1_ogg);
+				FileOutputStream out = new FileOutputStream(getFilesDir() + "/loop1_ogg.ogg");
+				byte[] buffer = new byte[1024];
+				int len;
+				while ((len = loop1.read(buffer)) != -1)
+					out.write(buffer, 0, len);
+				loop1.close();
+			}
+		} catch (FileNotFoundException e) {
+			Log.e("mp3decoders", "FileNotFoundException", e);
+		} catch (IOException e) {
+			Log.e("mp3decoders", "IOException", e);
 		}
 	}
